Introduced the new reproductive system very buggy

+++ src/world/world_settings.h
@@ -7,7 +7,7 @@ struct WorldSettings
 
 	inline static uint32_t cells_x;
 	inline static uint32_t cells_y;
-	inline static uint32_t cell_max_capacity;
+	inline static uint8_t cell_max_capacity;
 
 	inline static float border_repulsion_magnitude; // how strong it is repelled from the border
 	inline static unsigned int updating_threads = 16;
